refactored testing architecture. Added nested directory structure with CMakeLists for finer control of which tests to build. Also added readmes in each subdirectory
