+++ src/model/processing/Model.h
@@ -6,7 +6,6 @@
 
 #include "CloudType.h"
 #include <Eigen/Dense>
-#include <pcl/io/pcd_io.h>
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
 #include <memory>
