+++ test/CMakeLists.txt
@@ -8,19 +8,18 @@ include_directories(
         ${gtest_SOURCE_DIR}/include ${gtest_SOURCE_DIR}
         ${gmock_SOURCE_DIR}/include ${gmock_SOURCE_DIR}
 )
-file(GLOB TEST_SRC_FILES
-        ${swag_scanner_SOURCE_DIR}/test/*.cpp
-        ${swag_scanner_SOURCE_DIR}/test/calibrationTests/*.cpp
-        ${swag_scanner_SOURCE_DIR}/test/equationsTests/*.cpp
-        ${swag_scanner_SOURCE_DIR}/test/fileTests/*.cpp
-        ${swag_scanner_SOURCE_DIR}/test/processingTests/*.cpp
-        ${swag_scanner_SOURCE_DIR}/test/typesTests/*.cpp
-        ${swag_scanner_SOURCE_DIR}/test/utilsTests/*.cpp
-        )
 
-add_executable(${TEST_MAIN} ${TEST_SRC_FILES})
+add_executable(${TEST_MAIN} main.cpp)
 target_link_libraries(${TEST_MAIN} PRIVATE swag_scanner_lib
         gtest
         gtest_main)
+
+add_subdirectory(calibrationTests)
+add_subdirectory(equationsTests)
+add_subdirectory(fileTests)
+add_subdirectory(processingTests)
+add_subdirectory(research)
+add_subdirectory(typesTests)
+add_subdirectory(utilsTests)
 
 add_test(${TEST_MAIN} ${TEST_MAIN})
