+++ test/calibrationTests/CalibrationTests.cpp
@@ -1,20 +1,16 @@
 #include <gtest/gtest.h>
-#include <pcl/point_cloud.h>
-#include <pcl/point_types.h>
 #include <memory>
 #include "Model.h"
 #include "Visualizer.h"
-#include "Segmentation.h"
-#include "Algorithms.h"
 #include "Calibration.h"
 #include "Point.h"
+
 
 
 class CalibrationPhysicalFixture : public ::testing::Test {
 
 protected:
     model::Model *mod;
-    std::string test_folder_path = "/Users/seanngpack/Programming Stuff/Projects/scanner_files/testing/FileHandlerPhysicalTests";
     equations::Normal g_n = equations::Normal(-0.0158, -0.8661, -0.4996);
     std::vector<equations::Plane> planes = {
             equations::Plane(0.8603, -0.2446, 0.4472, -.201376),
@@ -71,37 +67,4 @@ TEST_F(CalibrationPhysicalFixture, TestBuildbMatrix) {
     ASSERT_NEAR(b(0, 0), -0.0414, .001);
     ASSERT_NEAR(b(5, 0), -0.0117, .001);
     ASSERT_NEAR(b(8, 0), 0.0153, .001);
-}
-
-TEST_F(CalibrationPhysicalFixture, VisualizePlaneCalibration) {
-
-    std::string folder_path = "/Users/seanngpack/Programming Stuff/Projects/scanner_files/18";
-    auto cloudIn = std::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
-    auto cloudOut = std::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
-    auto transformed = std::make_shared<pcl::PointCloud<pcl::PointXYZ>>();
-    pcl::io::loadPCDFile<pcl::PointXYZ>(folder_path + "/filtered/" + "1.pcd", *cloudIn);
-    pcl::io::loadPCDFile<pcl::PointXYZ>(folder_path + "/filtered/" + "4.pcd", *cloudOut);
-//    std::vector<float> coefs = mod->get_plane_coefs(cloudIn);
-//    mod->get_plane_coefs(cloudOut);
-
-    std::vector<float> origin = {-0.000213082, 0.0298714, 0.42673};
-    std::vector<float> direction = {-0.0158, -0.8661, -0.4996};
-
-    // from 1.pcd -> 4.pcd
-    float theta = 0.15708;
-
-    transformed->resize(cloudIn->size());
-
-    for (int i = 0; i < cloudIn->size(); i++) {
-        transformed->points[i] = algos::rotate_point_about_line(cloudIn->points[i],
-                                                                origin,
-                                                                direction,
-                                                                theta);
-    }
-
-
-//    visual::Visualizer visualizer;
-//    std::vector<std::shared_ptr<pcl::PointCloud<pcl::PointXYZ>>> clouds{cloudOut, transformed};
-//    visualizer.simpleVis(clouds);
-
 }
