+++ test/calibrationTests/README.md
@@ -1,0 +1,7 @@
+#calibrationTests
+
+This folder contains tests for verifying mathematical and visual accuracy
+of the calibration algorithm.
+
+* [CalibrationTests.cpp](./CalibrationTests.cpp) : Verifies building the calibration matrices and center point calculation
+* [CalibrationTestsVisual.cpp](visual/CalibrationTestsVisual.cpp) : Visual tests
