+++ test/equationsTests/README.md
@@ -1,0 +1,7 @@
+#equationsTests
+
+This folder contains tests for verifying custom equations:: types.
+
+* [EquationsTests.cpp](./EquationsTests.cpp) : 
+* [NormalTests.cpp](./NormalTests.cpp) : 
+* [PlaneTests.cpp](./PlaneTests.cpp) :
