+++ test/fileTests/README.md
@@ -1,0 +1,7 @@
+#fileTests
+
+This folder contains tests for verifying file handler behavior. Most of these tests
+depend on having infrastructure on your computer set up.
+
+* [ScanFileHandlerPhysicalTests.cpp](./ScanFileHandlerPhysicalTests.cpp) : Verifies Scan file handler constructors
+* [ScanFileHandlerTests.cpp](./ScanFileHandlerTests.cpp) : Verifies the handler creates folders and files correctly
