+++ test/processingTests/README.md
@@ -1,0 +1,11 @@
+#processingTests
+
+This folder contains tests for verifying mathematical and visual accuracy
+of processing algorithms.
+
+* [DepthTests.cpp](./DepthTests.cpp) : Verifies depth related methods such as creating pointclouds with depth frames.
+* [ModelTests.cpp](./ModelTests.cpp) : Verifies model methods
+* [ModelTests.cpp](./ModelTests.cpp) : Verifies model methods
+* [RegistrationPhysicalTests.cpp](./RegistrationPhysicalTests.cpp) : Verifies registration methods using premade example files in folder
+* [ModelTestsVisual.cpp](./visual/ModelTestsVisual.cpp) : Verifies model methods visually
+* [ModelTestsVisual.cpp](./visual/SegmentationTestsVisual.cpp) : Verifies segmentation methods visually
