+++ test/research/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(${TEST_MAIN} PRIVATE
+        ${CMAKE_CURRENT_SOURCE_DIR}/QualitativeComparisons.cpp
+        )
+
+target_include_directories(${TEST_MAIN} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
