+++ test/research/README.md
@@ -1,0 +1,3 @@
+#calibrationTests
+
+This folder contains tests and experiments for improving scan quality.
