+++ test/typesTests/README.md
@@ -1,0 +1,6 @@
+#typesTests
+
+This folder contains tests for verifying created types.
+
+* [CameraTypesTests.cpp](./CameraTypesTests.cpp) : Verifies CameraTypes
+* [CameraTypesTests.cpp](./CameraTypesTests.cpp) : Visual CloudTypes
