+++ test/utilsTests/README.md
@@ -1,0 +1,5 @@
+#utilsTests
+
+This folder contains tests for verifying different utility classes.
+
+* [AlgosTests.cpp](./AlgosTests.cpp) : Verifies mathematical and functional accuracy of handmade algorithms
